requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.8
eralchemy2==1.4.1
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.2.0
packaging==24.2
psycopg2-binary==2.9.10
pygraphviz==1.14
python-dotenv==1.0.0
PyYAML==6.0.2
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.0.1
